3° Save

+++ demo/src/main/java/com/ipss/demo/repository/MesaRepository.java
@@ -3,4 +3,5 @@ package com.ipss.demo.repository;
 import com.ipss.demo.model.Mesa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MesaRepository extends JpaRepository<Mesa, Long> {}
+public interface MesaRepository extends JpaRepository<Mesa, Integer> { }
+
